🐙️ueli🎆️26🌙️06☀️04🚩️578
🎆️26🌙️08☀️28⏰️00⌚️10⏱️37
🧪️Add repo test domain suite with multi-language test contract, discover, and doctor gates
🤖️Add GitHub Actions workflow for repo test domain CI validation
⚖️Add test domain inspection and validation gates to VS Code launch configurations
🦀️Add Rust divergence callback and writable path authority test cases
🧬️Add taxonomy leading grapheme and pattern compiler reuse test suites
🧹️Add normalization source admission test fixtures and schema definitions
📓️Add semantic mutation overhaul architectural plans, reviews, and audits

📊️metric📃uloc💯️3.62M📈️188k➗️5.49➕️191k✏️7.8k➖️2.86k🟰️202k
📊️metric💾size💯️244MB📈️69.8MB➗️40➕️70.2MB✏️23.4MB➖️375KB🟰️94MB
📊️metric🦀️rust📃uloc💯️2.01M📈️19k➗️0.955➕️21.4k✏️3.82k➖️2.36k🟰️27.6k
📊️metric🦀️rust💾size💯️108MB📈️1.07MB➗️1➕️1.21MB✏️10.6MB➖️140KB🟰️11.9MB
📊️metric🧾️json📃uloc💯️970k📈️139k➗️16.7➕️139k✏️2.74k➖️238🟰️142k
📊️metric🧾️json💾size💯️105MB📈️66.3MB➗️170➕️66.6MB✏️4.13MB➖️224KB🟰️70.9MB
📊️metric🟦️typescript📃uloc💯️389k📈️29.5k➗️8.19➕️29.7k✏️1.23k➖️258🟰️31.2k
📊️metric🟦️typescript💾size💯️21.7MB📈️2.41MB➗️12.5➕️2.42MB✏️8.7MB➖️12KB🟰️11.1MB
📊️metric🔵️go📃uloc💯️84.4k
📊️metric🔵️go💾size💯️2.81MB
📊️metric🟨️javascript📃uloc💯️55.6k
📊️metric🟨️javascript💾size💯️2.18MB
📊️metric🐍️python📃uloc💯️28.2k
📊️metric🐍️python💾size💯️1.37MB
📊️metric🎨️css📃uloc💯️27k
📊️metric🎨️css💾size💯️684KB
📊️metric📝️markdown📃uloc💯️25.7k📈️339➗️1.34➕️339✏️12🟰️351
📊️metric📝️markdown💾size💯️791KB📈️19.1KB➗️2.48➕️19.1KB✏️7.2KB🟰️26.3KB
📊️metric📐️tex📃uloc💯️9.45k
📊️metric📐️tex💾size💯️473KB
📊️metric⚙️toml📃uloc💯️6.24k📈️151➗️2.48➕️151🟰️151
📊️metric⚙️toml💾size💯️302KB📈️6.04KB➗️2.04➕️6.04KB✏️26.8KB🟰️32.9KB
📊️metric🌐️html📃uloc💯️4.93k
📊️metric🌐️html💾size💯️305KB
📊️metric📄️xml📃uloc💯️1.36k
📊️metric📄️xml💾size💯️135KB
📊️metric🛢️sql📃uloc💯️1.22k
📊️metric🛢️sql💾size💯️57.4KB
📊️metric💠️powershell📃uloc💯️906
📊️metric💠️powershell💾size💯️37.1KB
📊️metric🐚️shell📃uloc💯️604
📊️metric🐚️shell💾size💯️22.1KB
📊️metric🟣️csharp📃uloc💯️507
📊️metric🟣️csharp💾size💯️23KB
📊️metric📑️csv📃uloc💯️160
📊️metric📑️csv💾size💯️28.5KB
📊️metric🐳️dockerfile📃uloc💯️43
📊️metric🐳️dockerfile💾size💯️991B

+++ 🧰️framework/🔨️modules/🖱️ui/🧬️contract/♻️retirement/📮️handback/🦀️component.rs
@@ -1,13 +1,13 @@
 //! 📮️ Fixed arena-slot handback obligations; producers never acquire the arena mutex.
 
-use std::sync::atomic::{AtomicBool, AtomicU64, AtomicUsize, Ordering};
+use std::sync::atomic::{AtomicBool, AtomicU64, Ordering};
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub(crate) enum UiArenaHandback { ReleaseAlias, ReturnClaim }
 
 //#region 📫️FixedObligations
 pub(crate) struct UiArenaHandbacks<const SLOTS: usize, const WORDS: usize> {
-    releases: [AtomicUsize; SLOTS],
+    releases: [AtomicU64; SLOTS],
     claims: [AtomicBool; SLOTS],
     ready: [AtomicU64; WORDS],
 }
@@ -15,13 +15,13 @@ pub(crate) struct UiArenaHandbacks<const SLOTS: usize, const WORDS: usize> {
 impl<const SLOTS: usize, const WORDS: usize> UiArenaHandbacks<SLOTS, WORDS> {
     pub(crate) const fn new() -> Self {
         assert!(WORDS * 64 >= SLOTS);
-        Self { releases: [const { AtomicUsize::new(0) }; SLOTS], claims: [const { AtomicBool::new(false) }; SLOTS], ready: [const { AtomicU64::new(0) }; WORDS] }
+        Self { releases: [const { AtomicU64::new(0) }; SLOTS], claims: [const { AtomicBool::new(false) }; SLOTS], ready: [const { AtomicU64::new(0) }; WORDS] }
     }
 
     /// 📨️ Records an admitted owner; slot reuse remains forbidden until its obligation is consumed.
     pub(crate) fn record(&self, slot: usize, obligation: UiArenaHandback) {
         match obligation {
-            UiArenaHandback::ReleaseAlias => { let previous = self.releases[slot].fetch_add(1, Ordering::AcqRel); assert!(previous != usize::MAX, "pending releases exceed admitted aliases"); }
+            UiArenaHandback::ReleaseAlias => { let previous = self.releases[slot].fetch_add(1, Ordering::AcqRel); assert!(previous != u64::MAX, "pending releases exceed admitted aliases"); }
             UiArenaHandback::ReturnClaim => { assert!(!self.claims[slot].swap(true, Ordering::AcqRel), "one exact claim returned twice"); }
         }
         self.ready[slot / 64].fetch_or(1u64 << (slot % 64), Ordering::Release);
